+++ server.js
@@ -1,14 +1,15 @@
-// const express = require("express");
-// const path = require("path");
-// const exphbs = require("express-handlebars");
-// // const scripts = [{ logger: "./src/utils/logger" }];
-// const message = require('./src/data');
-// const profile = require('./src/formData');
+const express = require("express");
+const path = require("path");
+const exphbs = require("express-handlebars");
+// const scripts = [{ logger: "./src/utils/logger" }];
+const message = require('./src/data');
+const profile = require('./src/formData');
 
-// const app = express();
-// const PORT = 3000;
+const app = express();
+const PORT = 3000;
 
-// app.use(express.static(__dirname + ""));
+app.use(express.static(__dirname + ""));
+// app.set("view engine", "hbs");
 
 // app.engine(
 //     "hbs",
@@ -18,51 +19,51 @@
 //     })
 // );
 
-// app.set("view engine", "hbs");
-// app.set("views", path.join(__dirname, "views"));
+app.set("view engine", "hbs");
+app.set("views", path.join(__dirname, "views"));
 
-// app.get("/", (req, res) => {
-//     res.redirect("/login");
-//   });
+app.get("/", (req, res) => {
+    res.render("/login");
+  });
 
-// app.get("/about", (req, res) => {
-//     res.render("about");
-// });
+app.get("/about", (req, res) => {
+    res.render("about");
+});
 
-// app.get("/login", (req, res) => {
-//   res.render("login");
-// });
+app.get("/login", (req, res) => {
+  res.render("login");
+});
 
-// app.get("/signup", (req, res) => {
-//   res.render("signup");
-// });
+app.get("/signup", (req, res) => {
+  res.render("signup");
+});
 
-// app.get("/500", (req, res) => {
-//   res.render("500");
-// });
+app.get("/500", (req, res) => {
+  res.render("500");
+});
 
-// app.get("/404", (req, res) => {
-//   res.render("404");
-// });
+app.get("/404", (req, res) => {
+  res.render("404");
+});
 
-// app.get("/profile", (req, res) => {
-//     res.render("profile", profile);
-// });
+app.get("/profile", (req, res) => {
+    res.render("profile", profile);
+});
 
-// app.get("/chat", (req, res) => {
-//     res.render("chat", message );
-// });
+app.get("/chat", (req, res) => {
+    res.render("chat", message );
+});
 
-// app.listen(PORT, function () {
-//     console.log(`Example app listening on port ${PORT}!`);
-// });
+app.listen(PORT, function () {
+    console.log(`Example app listening on port ${PORT}!`);
+});
 
-// module.exports = app;
-const http = require('http')
-const app = require('./app')
+module.exports = app;
+// const http = require('http')
+// const app = require('./app')
 
-const port = process.env.PORT || 3000
+// const port = process.env.PORT || 3000
 
-const server = http.createServer(app)
+// const server = http.createServer(app)
 
-server.listen(port)
+// server.listen(port)
