+++ views/chat.hbs
@@ -1,31 +1,46 @@
-<div class="page-container">
-    <div class="left-menu-panel">
-        <div class="container">
-            <p class="message-top-line">
-                <input
-                    class="login-input"
-                    type="text"
-                    id="searcbar"
-                    placeholder="поиск"
-                    name="message-search"
-                />
-                <a href="/profile">Профиль</a>
-            </p>
-        </div>
-        {{#each message}}
-            <div class="container">
-                <img src={{imgSrc}} alt="Avatar" style="width:90px" />
-                <p class="message-top-line"><span
-                        class="message-name"
-                    >{{name}}</span><span
-                        class="message-time"
-                    >{{time}}</span></p>
-                <p class="ellipsis">{{text}}</p>
-            </div>
-        {{/each}}
+<html lang='en'>
+    <head>
+        <link rel='stylesheet/less' type='text/css' href='../../style.less' />
+        <title>Chat</title>
+        <script src='//cdn.jsdelivr.net/npm/less@3.13'></script>
+        <script
+            src='https://cdn.jsdelivr.net/npm/handlebars@latest/dist/handlebars.js'
+        ></script>
+        <script type='module' src='../../src/utils/logger.js'></script>
+    </head>
 
-    </div>
-    <div class="dialog-panel">
-        <div class="message-panel">Выберите чат чтобы отправить сообщение</div>
-    </div>
-</div>
+    <body>
+        <div class='page-container'>
+            <div class='left-menu-panel'>
+                <div class='container'>
+                    <p class='message-top-line'>
+                        <input
+                            class='login-input'
+                            type='text'
+                            id='searcbar'
+                            placeholder='поиск'
+                            name='message-search'
+                        />
+                        <a href='/profile'>Профиль</a>
+                    </p>
+                </div>
+                {{#each message}}
+                    <div class='container'>
+                        <img src={{imgSrc}} alt='Avatar' style='width:90px' />
+                        <p class='message-top-line'><span
+                                class='message-name'
+                            >{{name}}</span><span
+                                class='message-time'
+                            >{{time}}</span></p>
+                        <p class='ellipsis'>{{text}}</p>
+                    </div>
+                {{/each}}
+
+            </div>
+            <div class='dialog-panel'>
+                <div class='message-panel'>Выберите чат чтобы отправить
+                    сообщение</div>
+            </div>
+        </div>
+    </body>
+</html>
