+++ views/layouts/main.hbs
@@ -7,11 +7,6 @@
             src="https://cdn.jsdelivr.net/npm/handlebars@latest/dist/handlebars.js"
         ></script>
         <script type="module" src="../../src/utils/logger.js"></script>
-        {{!-- {{#each scripts}}
-            <script src="{{script}}"></script>
-        {{/each}} --}}
-
-        {{! <script type="module" src="../src/data.js"></script> }}
     </head>
 
     <body>
