+++ views/login.hbs
@@ -1,22 +1,38 @@
-<div class="profile-page-container">
-    <div class="profile-page-content login-wrap">
-        <span class="login-title">Вход</span>
-        {{! <form> }}
-        <input
-            type="text"
-            placeholder="логин"
-            name="login-enter"
-            id="login"
-            class="login-input"
-        />
-        <input
-            type="password"
-            placeholder="пароль"
-            name="password-enter"
-            class="login-input"
-        />
-        <button class="login-button login-button-margin"><a href="/chat">Авторизоваться</a></button>
-        <button class="register-button"><a href="/signup">Нет аккаунта?</a></button>
-        {{! </form> }}
-    </div>
-</div>
+<html lang='en'>
+    <head>
+        <link rel='stylesheet/less' type='text/css' href='../../style.less' />
+        <title>Chat</title>
+        <script src='//cdn.jsdelivr.net/npm/less@3.13'></script>
+        <script
+            src='https://cdn.jsdelivr.net/npm/handlebars@latest/dist/handlebars.js'
+        ></script>
+        <script type='module' src='../../src/utils/logger.js'></script>
+    </head>
+
+    <body>
+        <div class='profile-page-container'>
+            <div class='profile-page-content login-wrap'>
+                <span class='login-title'>Вход</span>
+                {{! <form> }}
+                <input
+                    type='text'
+                    placeholder='логин'
+                    name='login-enter'
+                    id='login'
+                    class='login-input'
+                />
+                <input
+                    type='password'
+                    placeholder='пароль'
+                    name='password-enter'
+                    class='login-input'
+                />
+                <button class='login-button login-button-margin'><a
+                        href='/chat'
+                    >Авторизоваться</a></button>
+                <button class='register-button'><a href='/signup'>Нет аккаунта?</a></button>
+                {{! </form> }}
+            </div>
+        </div>
+    </body>
+</html>
